tourguider/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.5
django-filter==1.0.4
django-rest-swagger==2.1.2
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
drfdocs==0.0.11
factory-boy==2.9.2
Faker==0.8.4
graphene==2.0.dev20170802065539
graphene-django==2.0.dev2017083101
graphene-django-extras==0.0.1b10
graphql-core==2.0.dev20170801051721
graphql-relay==0.4.5
idna==2.6
iso8601==0.1.12
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
olefile==0.44
openapi-codec==1.3.2
Pillow==4.2.1
promise==2.0.2
PyJWT==1.5.3
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
simplejson==3.11.1
singledispatch==3.4.0.3
six==1.11.0
snapshottest==0.5.0
termcolor==1.1.0
typing==3.6.2
uritemplate==3.0.0
urllib3==1.22
